+++ govtool/frontend/src/components/molecules/GovernanceActionCard.tsx
@@ -54,11 +54,11 @@ export const GovernanceActionCard: FC<ActionTypeProps> = ({ ...props }) => {
   const { t } = useTranslation();
 
   const govActionId = getFullGovActionId(txHash, index);
-  const cip129GovernanceActionId = encodeCIP129Identifier(
-    txHash,
-    index.toString(16).padStart(2, "0"),
-    "gov_action",
-  );
+  const cip129GovernanceActionId = encodeCIP129Identifier({
+    txID: txHash,
+    index: index.toString(16).padStart(2, "0"),
+    bech32Prefix: "gov_action",
+  });
 
   return (
     <Box